Python/prompt-sdk/pyproject.toml
[tool.poetry]
name = "prompt-sdk"
description = "Safely integrate prompts managed in the Autoblocks platform into your application"
version = "0.0.0"
authors = ["Autoblocks Engineering <engineering@autoblocks.ai>"]
license = "MIT"
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
openai = "^1.0.0"
autoblocksai = "^0.0.39"
python-dotenv = "^1.0.0"
pydantic = "^2.5.3"

[tool.poetry.scripts]
start = "src.main:main"
